separate param function in separate file
In addition:
-change "gpu" keyword into "emulate" keyword (consistency)
-destroy context on window close
-remove try..excpt from loading gpu.gpu.py
